requirements.txt
# FastAPI Gateway
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2

# Detection
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.2

# Redis Queue
redis==5.0.1
rq==1.16.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0

# Image Processing
Pillow==10.1.0
pytesseract==0.3.10

# Audio Processing
openai-whisper==20231117
soundfile==0.12.1

# Utilities
pyyaml==6.0.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Alerts
slack-sdk==3.26.1

# Analytics
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
